+++ CargoVessel.cpp
@@ -1,3 +1,4 @@
+#pragma once
 #include "CargoVessel.h"
 
 DataStruct CargoVessel::GetDataStruct()
@@ -18,8 +19,7 @@ bool CargoVessel::UnpackCargo(bool displayMessage)
 	status.version = 2;
 	GetStatusEx(&status);
 
-	SetSpawnName();
-	OBJHANDLE cargo = oapiCreateVesselEx(dataStruct.spawnName.c_str(), dataStruct.spawnModule.c_str(), &status);
+	OBJHANDLE cargo = oapiCreateVesselEx(SetSpawnName().c_str(), dataStruct.spawnModule.c_str(), &status);
 
 	if (cargo) {
 		if (displayMessage) DisplayMessage("cargo unpacked successfully");
@@ -41,21 +41,22 @@ double CargoVessel::UseResource(double requiredMass)
 	return drainedMass;
 }
 
-void CargoVessel::SetSpawnName()
+std::string CargoVessel::SetSpawnName()
 {
-	if (!oapiGetVesselByName(&dataStruct.spawnName[0])) return;
+	if (!oapiGetVesselByName(&dataStruct.spawnName[0])) return dataStruct.spawnName;
 
 	for (int index = 1; index++;)
 	{
 		std::string spawnName = dataStruct.spawnName.c_str() + std::to_string(index - 1);
- 		if (!oapiGetVesselByName(&spawnName[0])) { dataStruct.spawnName = spawnName; return; }
+ 		if (!oapiGetVesselByName(&spawnName[0])) return spawnName;
 	}
+
+	return std::string();
 }
 
 void CargoVessel::DisplayMessage(std::string message)
 {
 	std::string cargoName = GetName();
-	cargoName += " ";
-	cargoName += message;
+	cargoName += " " + message;
 	sprintf(oapiDebugString(), cargoName.c_str());
 }