+++ SDK/UCSO_SDK.h
@@ -1,137 +1,138 @@
+// ==============================================================
+// UCSO_SDK.h : Defines the Universal Cargo System for Orbiter public SDK.
+// 
+// Copyright © 2019 Abdullah Radwan
+// All rights reserved.
+//
+// ==============================================================
+
+#pragma once
 #include <orbitersdk.h>
 #include <map>
 #include "../CargoVessel.h"
 
-#pragma comment(lib,"UCSO_SDK.lib")
 class UCSO
 {
 public:
+	// The grapple result, as returned from GrappleCargo function.
 	enum GrappleResult {
-		CARGO_GRAPPLED = 0, /** Cargo is grappled successfully. */
-		NO_CARGO_IN_RANGE, /** No cargo in range. */
-		MAX_MASS_EXCEEDED, /** Maximum one cargo mass will be exceeded if the cargo is added. */
-		MAX_TOTAL_MASS_EXCEEDED, /** Maximum total cargo mass will be exceeded if the cargo is added. */
-		SLOT_OCCUPIED,     /** The given slot is occupied. */
-		SLOT_UNDEFINED,    /** The given slot is undefiend. */
-		GRAPPLE_FAILED     /** Grapple has failed. */
+		CARGO_GRAPPLED = 0,      // The cargo is grappled successfully.
+		NO_CARGO_IN_RANGE,       // No cargo in the grapple range.
+		MAX_MASS_EXCEEDED,       // The maximum one cargo mass will be exceeded if the cargo is added.
+		MAX_TOTAL_MASS_EXCEEDED, // The maximum total cargo mass will be exceeded if the cargo is added.
+		SLOT_OCCUPIED,           // The passed slot is occupied.
+		SLOT_UNDEFINED,          // The passed slot is undefiend.
+		GRAPPLE_FAILED           // The grapple failed.
 	};
+
+	// The release result, as returned from ReleaseCargo, UnpackCargo, and DeleteCargo functions.
 	enum ReleaseResult
 	{
-		CARGO_RELEASED = 0, /** Cargo is released successfully. */
-		SLOT_EMPTY,        /** The given slot is empty, or all slots are empty if -1 is passed. */
-		SLOT_UNDEF,        /** The given slot is undefined, or the no slot is defined. */
-		RELEASE_FAILED     /** Release has failed. */
+		CARGO_RELEASED = 0, // The cargo is released successfully. For ReleaseCargo function only.
+		SLOT_EMPTY,         // The passed slot is empty, or all slots are empty if -1 is passed.
+		SLOT_UNDEF,         // The passed slot is undefined, or the no slots are defined.
+		RELEASE_FAILED      // The release failed. For ReleaseCargo function only.
 	};
+
+	// The unpack result, as returned from UnpackCargo function.
 	enum UnpackResult
 	{
-		CARGO_UNPACKED = 4, /** Cargo is unpacked successfully. */
-		NOT_UNPACKABLE, /** Cargo is unpackable. */
-		UNPACK_FAILED  /** Unpack has failed. */
+		CARGO_UNPACKED = 4, // The cargo is unpacked successfully.
+		NOT_UNPACKABLE,     // The cargo is unpackable.
+		UNPACK_FAILED       // The unpack failed.
 	};
+
+	// The delete result as returned from DeleteCargo function.
 	enum DeleteResult
 	{
-		CARGO_DELETED = 4, /** Cargo is deleted successfully. */
-		DELETE_FAILED      /** Delete has failed. */
+		CARGO_DELETED = 4, // The cargo is deleted successfully.
+		DELETE_FAILED      // The delete failed.
 	};
 
-	/**
-	 * \brief Create UCSO instance.
-	 * \param vessel pointer to the vessel.
-	 * \return a UCSO instance.
-	*/
-	static UCSO* Init(VESSEL* vessel);
+	// Performs one-time initialization of UCSO. It should be called from your vessel's constructor.
+	// Parameters:
+	//	vessel: pointer to the calling vessel.
+	// Returns a UCSO instance.
+	// NOTE: Don't forget to delete the returned object when you no longer need it (e.g., in your vessel's destructor).
+	static UCSO* CreateInstance(VESSEL* vessel);
 
-	/**
-	 * \brief Set the slot number for a given attachment, to use with the SDK functions.
-	 * \param slot The slot number.
-     * You can use any number but not -1, as it's reserved.
-	 * You can use the number multiple times to update the slot attachment handle.
-	 * \param attachmentHandle The attachment handle.
-	*/
+	// Sets the slot number for a given attachment, to use with the SDK functions. It must be called before any other calls to the SDK.
+	// To delete a slot, pass the slot number and NULL for the attachment handle.
+	// Parameters:
+	//	slot: the slot number. You can use any number but not -1, as it's reserved.
+	//		You can use the number multiple times to update the slot attachment handle.
+	//	attachmentHandle: the attachment handle. If NULL, the slot will be removed.
 	void SetSlotAttachment(int slot, ATTACHMENTHANDLE attachmentHandle);
 
-	/**
-	 * \brief Set the maximum one cargo mass for the vessel.
-	 * If any cargo mass exceeded this mass, the cargo won't be added.
-	 * \param maxCargoMass the maximum mass in kilograms. Set -1 for unlimited mass. The default value is -1.
-	*/
+	// Sets the maximum one cargo mass for the vessel.
+	// If any cargo mass exceeded this mass, the cargo won't be added.
+	// Parameters:
+	//	maxCargoMass: the maximum mass in kilograms. Set -1 for unlimited mass. The default value is -1.
 	void SetMaxCargoMass(double maxCargoMass);
 
-	/**
-	 * \brief Set the maximum total cargo mass for the vessel.
-	 * If the total cargo mass will exceed this mass if the cargo is added, the cargo won't be added.
-	 * \param maxTotalCargoMass the maximum mass in kilograms. Set -1 for unlimited mass. The default value is -1.
-	*/
+	// Sets the maximum total cargo mass for the vessel.
+	// If the total cargo mass will exceed this mass if the cargo is added, the cargo won't be added.
+	// Parameters:
+	//	maxTotalCargoMass: the maximum mass in kilograms. Set -1 for unlimited mass. The default value is -1.
 	void SetMaxTotalCargoMass(double maxTotalCargoMass);
 
-	/**
-	 * \brief Set the maximum grapple distance in meters.
-	 * \param grappleDistance the distance in meters. The default value is 50 meters.
-	*/
+	// Sets the maximum grapple distance in meters.
+	// Parameters:
+	//	grappleDistance: the distance in meters. The default value is 50 meters.
 	void SetMaxGrappleDistance(double grappleDistance);
 
-	/**
-	 * \brief Set the cargo release velocity if released in space.
-	 * \param releaseVel the release velocity in m/s. The default value is 0.05 m/s.
-	*/
+	// Sets the cargo release velocity if released in space.
+	// Parameters:
+	//	releaseVel: the release velocity in m/s. The default value is 0.05 m/s.
 	void SetReleaseVelocity(double releaseVel);
 
-	/**
-	 * \brief Grapple the cargo in the passed slot.
-	 * \param slot the slot number. If no slot is passed, the first empty slot will be used.
-	 * \return the result as the GrappleResult enum.
-	*/
+	// Grapples the cargo in the passed slot.
+	// Parameters:
+	//	slot: the slot number. If no slot is passed, the first empty slot will be used.
+	// Returns the result as the GrappleResult enum.
 	int GrappleCargo(int slot = -1);
 
-	/**
-	 * \brief Release the cargo in the passed slot.
-	 * \param slot the slot number. If no slot is passed, the first occupied slot will be used.
-	 * \return the result as the ReleaseResult enum.
-	*/
+	// Releases the cargo in the passed slot.
+	// Parameters:
+	//	slot: the slot number. If no slot is passed, the first occupied slot will be used.
+	// Returns the result as the ReleaseResult enum.
 	int ReleaseCargo(int slot = -1);
 
-	/**
-	 * \brief Unpack the cargo in the passed slot.
-	 * \param slot the slot number. If no slot is passed, the first occupied slot will be used.
-	 * \return the result as the GrappleResult and ReleaseResult enum.
-	*/
+	// Unpacks the cargo in the passed slot.
+	// Parameters:
+	//	slot the slot number. If no slot is passed, the first occupied slot will be used.
+	// Returns the result as the GrappleResult and ReleaseResult enum.
 	int UnpackCargo(int slot = -1);
 
-	/**
-	* \brief Delete the cargo in the passed slot.
-	* \param slot the slot number.
-	* \return the delete result as the DeleteResult and ReleaseResult enum.
-	*/
+	// Deletes the cargo in the passed slot.
+	// Parameters:
+	//	slot the slot number.
+	// Returns the delete result as the DeleteResult and ReleaseResult enum.
 	int DeleteCargo(int slot);
 
-	/**
-	  * \brief Use the available resource in the cargo in the passed slot.
-	  * \param resourceType the cargo type (see the standard cargo types in the manual).
-	  * \param requiredMass the needed mass in kilograms.
-      * \param slot the slot number. If no slot is passed, the first available cargo will be used.
-	  * \return the required mass or less based on the available mass, or 0 if:
-      * The cargo isn't a resource.
-	  * The cargo resource type doesn't match the passed type.
-	  * The cargo is empty.
-	  * The slot is empty or undefined.
-	*/
+	// Uses the available resource in the cargo in the passed slot.
+	// Parameters:
+	//	resourceType: the cargo type (see the standard cargo types in the manual).
+	//	requiredMass: the needed mass in kilograms.
+    //	slot: the slot number. If no slot is passed, the first available cargo will be used.
+	// Returns the required mass or less based on the available mass, or 0 if:
+    //	The cargo isn't a resource.
+	//	The cargo resource type doesn't match the passed type.
+	//	The cargo is empty.
+	//	The slot is empty or undefined.
 	double UseResource(std::string resourceType, double requiredMass, int slot = -1);
 
-	/**
-	 * \brief Get the cargo mass in the passed slot.
-	 * \param slot the slot number.
-	 * \return the cargo mass in kilograms, or -1 if the slot is empty or undefined.
-	*/
+	// Gets the cargo mass in the passed slot.
+	// Parameters:
+	//	slot: the slot number.
+	// Returns the cargo mass in kilograms, or -1 if the slot is empty or undefined.
 	double GetCargoMass(int slot);
 
-	/**
-	  * \brief Get the total cargo mass.
-	  * \return The total cargo mass in kilograms.
-	*/
+	// Gets the total cargo mass.
+	// Returns the total cargo mass in kilograms.
 	double GetCargoTotalMass();
 
 private:
-	UCSO() {};
 	UCSO(VESSEL* vessel);
 
 	VESSEL* vessel;
@@ -143,7 +144,7 @@ private:
 	double releaseVel;
 	double totalCargoMass;
 
-	bool isTotalMassGet = false;
+	bool isTotalMassGet;
 
 	OBJHANDLE VerifySlot(int slot);
 	int GetEmptySlot();
@@ -153,4 +154,3 @@ private:
 
 	CargoVessel* GetResourceCargo(std::string resourceType);
 };
-