+++ CHANGELOG.md
@@ -4,5 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Alpha 2 - 2019-08-14
+### Added
+- Remove a slot using SetSlotAttachment function.
+### Fixed
+- DetachChild doesn't apply the passed velocity (Thanks to Woo482).
+### Changed
+- Rename SDK function 'Init' to 'CreateInstance'.
+- The SDK documentation.
+- Use GroundContact function instead of GetFlightStatus to detect if the vessel is landed..
+- The Visual Studio solution to proper build settings.
+
 ## Alpha 1 - 2019-08-14
 Initial release.